+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0XPQAG2F60YWHK47T5HTSR8"
+        "01K10FCGXNGTDQ3Q9RV28J7K60"
     )
     r"""The id of the resource"""