+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-24T07:52:40.721Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-25T09:42:09.309Z")
     r"""The date and time the resource was last updated"""