+++ packages/orq-rc/RELEASES.md
@@ -2079,3 +2079,13 @@ Based on:
 - [python v3.10.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.18] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.18 - packages/orq-rc
+
+## 2025-07-25 09:43:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.589.0 (2.664.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.19] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.19 - packages/orq-rc